// packages/simplyui-react/src/components/indicator/indicator-dot/utils/get-indicator-dot-position-variables/get-indicator-dot-position-variables.types.ts
export type IndicatorDotPosition =
  | 'bottom-center'
  | 'bottom-end'
  | 'bottom-start'
  | 'middle-center'
  | 'middle-end'
  | 'middle-start'
  | 'top-center'
  | 'top-end'
  | 'top-start';

export type IndicatorDotPositionVariables =
  | '--indicator-dot-top'
  | '--indicator-dot-bottom'
  | '--indicator-dot-left'
  | '--indicator-dot-right'
  | '--indicator-dot-translate-x'
  | '--indicator-dot-translate-y';